+++ src/test/java/api/support/builders/HoldingRequestBuilder.java
@@ -10,22 +10,26 @@ public class HoldingRequestBuilder implements Builder {
   private final UUID instanceId;
   private final UUID permanentLocationId;
   private final UUID temporaryLocationId;
+  private final String callNumber;
 
   public HoldingRequestBuilder() {
     this(
       null,
       UUID.fromString(ApiTestSuite.getThirdFloorLocation()),
-      UUID.fromString(ApiTestSuite.getReadingRoomLocation()));
+      UUID.fromString(ApiTestSuite.getReadingRoomLocation()),
+      null);
   }
 
   private HoldingRequestBuilder(
     UUID instanceId,
     UUID permanentLocationId,
-    UUID temporaryLocationId) {
+    UUID temporaryLocationId,
+    String callNumber) {
 
     this.instanceId = instanceId;
     this.permanentLocationId = permanentLocationId;
     this.temporaryLocationId = temporaryLocationId;
+    this.callNumber = callNumber;
   }
 
   @Override
@@ -35,24 +39,29 @@ public class HoldingRequestBuilder implements Builder {
     holding.put("instanceId", instanceId.toString())
            .put("permanentLocationId", permanentLocationId.toString());
     
-    if (temporaryLocationId != null) 
+    if (temporaryLocationId != null) {
       holding.put("temporaryLocationId", temporaryLocationId.toString());
-    
+    }
+
+    holding.put("callNumber", callNumber);
+
     return holding;
   }
 
-  public HoldingRequestBuilder withPermanentLocation(UUID permanentLocationId) {
+  private HoldingRequestBuilder withPermanentLocation(UUID permanentLocationId) {
     return new HoldingRequestBuilder(
       this.instanceId,
       permanentLocationId,
-      this.temporaryLocationId);
+      this.temporaryLocationId,
+      callNumber);
   }
   
-  public HoldingRequestBuilder withTemporaryLocation(UUID temporaryLocationId) {
+  private HoldingRequestBuilder withTemporaryLocation(UUID temporaryLocationId) {
     return new HoldingRequestBuilder(
-            this.instanceId,
-            this.permanentLocationId,
-            temporaryLocationId);
+      this.instanceId,
+      this.permanentLocationId,
+      temporaryLocationId,
+      this.callNumber);
   }
     
   public HoldingRequestBuilder permanentlyInMainLibrary() {
@@ -71,6 +80,15 @@ public class HoldingRequestBuilder implements Builder {
     return new HoldingRequestBuilder(
       instanceId,
       this.permanentLocationId,
-      this.temporaryLocationId);
+      this.temporaryLocationId,
+      this.callNumber);
+  }
+
+  public HoldingRequestBuilder withCallNumber(String callNumber) {
+    return new HoldingRequestBuilder(
+      this.instanceId,
+      this.permanentLocationId,
+      this.temporaryLocationId,
+      callNumber);
   }
 }